strftime/gofmt: use correct days calculation
We incorrectly used the naiive daysSinceEpoch calculation to calculate
the weekday in strftime/gofmt. This resulted in disregarding some
specific timezone information and incorrect weekdays. Use the correct
calculation, which is daysFromCivil. This calculates the number of days
since the epoch from a civil date, rather than a timestamp. The civil
date is correct, and naiively calculating from the timestamp is not
correct in this case.

Add test for this specific case.

Fixes: #15