gofmt: require fractional seconds to be all 0 or 9
If a '.' or ',' is followed by a '0' or '9', we require that all
sequential digits be the same as the first. This allows us to parse
something like "02.01.2006" properly. The '.01' previously would be
parsed as a fractional second with precision 1. Now, it is parsed as a
'.' delimiter and a "01" as the next hunk.

Fixes: #27